Added draggable dice
Requested rolls and saves can now be dragged or double clicked to roll. Requested dice rolls or saves can be dragged to the dice tower to be hidden. Hidden rolls sent by the console are always hidden, even if dropped in the chat box. Popup throw button replaced with diecontrol reflecting relevant dice.